Some minimal quick-fixes for error handling:
When running in parallel mode (i.e. default), regression server errors or a persistently failing test request cause the processing pool to essentially freeze until the timeout, without running the rest of the tests.

There's _plenty_ more I'd like to change here to clean up flow/reporting etc, but for now, this is the minimal sufficient changes to be parallel-friendly & run as much of the test suite as possible.